Add handle_input flag for setting input event as handled
Added a handle_input flag that will determine if `get_tree().get_root().set_input_as_handled()` is called or not on input events. 

I added this because of an issue I had where I wanted a list of nodes inside the SmoothScrollContainer to have mouse enter and exit effects but the SmoothScrollContainer was forcing the events to be handled which prevents child nodes from receiving mouse input events.